Set the requested PropagationMode for volumes
We have disabled docker volumes in our Nomad clients to force users to have to
use host volumes since these are better controlled using ACL's.

For running cadvisor we need to expose the docker data directory to cadvisor, we
do this by allowing the cadvisor job to mount a host volume with the docker data
path and set propagationmode to "host-to-task" (rslave) as docker requires this,
see error:

```
failed to create container:
  API error (400):
    invalid mount config:
      must use either propagation mode "rslave" or "rshared" when mount source is within the daemon root
```

Currently this wasn't applied for Host nor CSI volume mounts and I don't know why.
This MR enables the option for both.