docs: backport fix for metrics on blocked_evals vs broker
In #15835 we renamed the `nomad.broker.total_blocked` metric to
`nomad.broker.total_pending`, but in the process identified that the existing
scheduling performance monitoring guide mixed up the `broker.total_blocked`
metric with the `blocked_evals.total_blocked` metric. This changeset backports
the fix to the docs without renaming the metric (for backwards compatibility).