test for worker backoff
Makes the `raftSyncLimit` into a parameter on the worker's `run` method so that
we can run backoff tests without waiting 30+ seconds. I haven't followed thru
and made all the worker globals into worker parameters, because there isn't much
use outside of testing, but we can consider that in the future.