scheduler: system scheduler should reconcile overlapping live allocs
The system scheduler tests have incorrect assertions for a test case that should
be impossible in the first place, which is when more than one running allocation
exists for a system job on the same node.

In this case, we expect one to be stopped and the other to be updated in
place (or destructively in the case of destructive updates), which was not
asserted by the tests. Unfortunately, this also appears to be a bug stemming
from how we use the name of the allocation (and not the ID) to index allocations
in the `diffSystemAllocs` function. How we get into the initial conditions is
not yet known.