api: Add TaskClient() convenience helper for tasks making use of Nomad API
This PR adds api.TaskClient(ConfigModifyFunc) helper method for use by
applications that are running as Nomad Tasks capable of making use of
Task API and workload identity. The resulting Client is automatically
properly configured to communicate with Nomad over the Task API unix
domain socket, as well as setting the necessary Authorization header.

Folks who need to further configure the Client may provide a ConfigModifyFunc
for modifying the underlying Config of the Client. This is done this
way because unlike with NewClient, in the 90% use case touching the
Config should not be necessary.