checksumming state store implementation
Nomad's in-memory state store is subject to corruption when an object that's
read from memdb is mutated. Depending on memdb's (undocumented) isolation
guarantees, there may be a narrow number of cases where an object that's been
inserted can be safely mutated in the same transaction before the transaction is
committed. But in general the rule is that you always need to deep copy the
object before mutating. This class of bug is hard to track down, leading to
"spooky action at a distance" when servers have diverging in-memory states, and
the potential existence of these bugs undermines our confidence in understanding
the system.

This changeset extends the FSM interface wrapper started by the event stream
work to allow checksumming. When enabled, the wrapper checksums each object on
insert and writes it to a table of checksums. On read, the wrapper checksums the
object again and looks for a matching previously-recorded checksum. If there's
no match, the object has been mutated outside of an insert.

The process of going through the whole code base to detect these problems is
going to land in several PRs, as we frequently corrupt the state store in test
code. Checksumming slows down the state store quite a bit (4x latency for
typical RPCs with multiple queries like `JobRegister`), so this will only ever
be intended for testing and developer debugging.