cli: refactor alloc reschedule with job restart
The previous implementation would only monitor the eval resulting from
the alloc stop action, but when multiple allocations for the same task
group are stopped at the same time (which can happen with `-batch-size`
> 1) and the replacement allocations failed to be placed only one of the
evals is blocked and include the failed metrics. The second eval is just
cancelled without any reference to the blocked eval, causing the
goroutine to be stuck waiting for a replacement alloc preventing the
whole command from exiting.

The new implementation monitors all evals for the job until a blocked
eval is found that contains failed placement metrics for the alloc task
group, or a replacement alloc succeeds.