client: refactor service interpolation
Services may reference runtime variables that must be interpolated in
the client when reading them, otherwise any equality check fails because
something like `${NOMAD_NAMESPACE}` is compared with its rendered value
stored in the service catalog.

This change attempts to make this requirement easier to discover by
providing methods in the `TaskGroup` and `Task` structs that return the
interpolated services.

These methods take a `ServiceInterpolator` interface to avoid a circular
dependency between the `structs` and `taskenv` packages.