cli: fix follow-up alloc monitoring
The `job restart` command monitors the allocation that was stopped and
its follow-up until it is running. But in some scenarios is may be
possible for the follow-up allocation itself to be replaced, such as in
case the new allocation fails.

In this scenario the command needs to keep following the
`NextAllocation` ID until it finds one that is running.