drainer: stop watching deleted, down, or disconnected nodes
When a node is down or disconnected, we can no longer gracefully migrate its
allocations. Any evaluations we need to replace the allocations will have
already been created by the heartbeater, so there's no more work for the drainer
to do. Stop watching nodes in this state.

Also, the blocking query for nodes set the maximum index to the highest index of
a node it found, rather than the index of the nodes table. This misses updates
to the index from deleting nodes. This was done as an performance optimization
to avoid excessive unblocking, but because the query is over all nodes anyways
there's no optimization to be had here. Remove the optimization so we can detect
deleted nodes without having to wait for an update to an unrelated node.

This changeset also refactors the tests of the draining node watcher so that we
don't mock the node watcher's `Remove` and `Update` methods for its own
tests. Instead we'll mock the node watcher's dependencies (the job watcher and
deadline notifier) and now unit tests can cover the real code. This allows us to
remove a bunch of TODOs in `watch_nodes.go` around testing.